filter_and_normalise_scRNA__fpkm

filters out poor quality cells and features, and converts counts matrix to FPKM

filter_and_normalise_scRNA__fpkm(
  counts_matrix,
  output_dir = "./",
  manual_filter = FALSE,
  filter_only = FALSE
)

Arguments

counts_matrix
a raw counts matrix from featureCounts

output_dir
output directory to save plots

manual_filter
vector of 4 values, manually specifying upper cutoffs to apply for total counts. e.g. c(20000,6000000,20,50) for upper limits of total counts, total_features_by_counts, pct mt, and pct ercc respectively. Also accepts lists instead of ints, where the first element of list is lower cutoff and second is upper cutoff.

filter_only
boolean to return only filtered (i.e. non-fpkm normalized) matrix

Value

FPKM normalized counts matrix